+++ sidebar-front-page.php
@@ -6,7 +6,7 @@
  * @copyright  WebMan Design, Oliver Juhas
  *
  * @since    1.0
- * @version  2.0
+ * @version  2.5.3
  */
 
 
@@ -33,7 +33,7 @@
 
 		<div id="front-page-widgets" class="front-page-widgets">
 
-			<aside id="front-page-widgets-container" class="widget-area front-page-widgets-container" role="complementary" aria-labelledby="sidebar-front-page-label">
+			<aside id="front-page-widgets-container" class="widget-area front-page-widgets-container" aria-labelledby="sidebar-front-page-label">
 
 				<h2 class="screen-reader-text" id="sidebar-front-page-label"><?php echo esc_attr_x( 'Front page sidebar', 'Sidebar aria label', 'auberge' ); ?></h2>