+++ template-parts/content-post-condensed.php
@@ -6,7 +6,7 @@
  * @copyright  WebMan Design, Oliver Juhas
  *
  * @since    2.0
- * @version  2.0
+ * @version  2.5.3
  */
 
 
@@ -17,7 +17,7 @@
 
 <?php do_action( 'tha_entry_before' ); ?>
 
-<article role="article" id="post-<?php the_ID(); ?>" <?php post_class(); echo apply_filters( 'wmhook_entry_container_atts', '' ); ?>>
+<article id="post-<?php the_ID(); ?>" <?php post_class(); echo apply_filters( 'wmhook_entry_container_atts', '' ); ?>>
 
 	<?php do_action( 'tha_entry_top' ); ?>